admin: factor out status routes
